+++ packages/client-auth/RELEASES.md
@@ -179,3 +179,13 @@ Based on:
 - [typescript v1.0.0-alpha.17] packages/client-auth
 ### Releases
 - [NPM v1.0.0-alpha.17] https://www.npmjs.com/package/@formance/sdk-auth/v/1.0.0-alpha.17 - packages/client-auth
+
+## 2024-08-02 00:09:09
+### Changes
+Based on:
+- OpenAPI Doc  
+- Speakeasy CLI 1.351.3 (2.384.4) https://github.com/speakeasy-api/speakeasy
+### Generated
+- [typescript v1.0.0-alpha.18] packages/client-auth
+### Releases
+- [NPM v1.0.0-alpha.18] https://www.npmjs.com/package/@formance/sdk-auth/v/1.0.0-alpha.18 - packages/client-auth